+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `DropdownSelectWithInput`: Fix selected item margin when dropdownDisabled.
+
 ## [3.17.0] - 2021-05-31
 
 Feature: